feature: move ChainUser and other client types to chain-api
chain-api module contains re-useable types and classes.
intended for use in client applications.
chain-client contains hyperledger fabric and other dependencies.
moving these types to chain-api allows easier re-use in some projects.
