+++ chain-api/src/client/api/ChainUserAPI.ts
@@ -12,8 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { UserAlias } from "@gala-chain/api";
-
+import { UserAlias } from "../../types";
 import { ChainClient, ChainUser } from "../generic";
 
 export interface ChainUserAPI {
