+++ chain-api/src/client/generic/ChainUser.spec.ts
@@ -12,8 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { signatures } from "@gala-chain/api";
-
+import { signatures } from "../../utils";
 import { ChainUser } from "./ChainUser";
 
 describe("ChainUser", () => {
