+++ chain-api/src/client/generic/ChainUser.ts
@@ -12,7 +12,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { UserAlias, signatures } from "@gala-chain/api";
+import { UserAlias } from "../../types";
+import { signatures } from "../../utils";
 
 /**
  * Represents a user configuration object, containing all the information
