+++ chain-api/src/client/generic/ContractConfig.ts
