+++ chain-api/src/client/generic/index.ts
@@ -12,4 +12,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from "./loadJson";
+
+export { ChainClient, isClassType, ClassType } from "./ChainClient";
+export { ChainClientBuilder } from "./ChainClientBuilder";
+export { ChainUser } from "./ChainUser";
+export { ContractConfig } from "./ContractConfig";
