+++ chain-api/src/client/index.ts
@@ -12,4 +12,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from "./loadJson";
+
+export * from "./api";
+export * from "./generic";
