+++ chain-cli/chaincode-template/e2e/lockNFT.spec.ts
@@ -14,6 +14,7 @@
  */
 import {
   AllowanceType,
+  ChainUser,
   GrantAllowanceDto,
   LockTokensDto,
   TokenAllowance,
@@ -23,7 +24,6 @@ import {
   UnlockTokensDto,
   createValidSubmitDTO
 } from "@gala-chain/api";
-import { ChainUser } from "@gala-chain/client";
 import {
   AdminChainClients,
   TestClients,
