+++ chain-client/src/hf/HFClientBuilder.ts
@@ -12,7 +12,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ChainClientBuilder, ContractConfig } from "../generic";
+import { ChainClientBuilder, ContractConfig } from "@gala-chain/api";
+
 import { CAClient } from "./CAClient";
 import { HFClient } from "./HFClient";
