+++ chain-client/src/index.ts
@@ -13,8 +13,6 @@
  * limitations under the License.
  */
 
-export * from "./api";
 export * from "./gcclient";
-export * from "./generic";
 export * from "./hf";
 export * from "./rest-api";
