+++ chain-client/src/rest-api/GlobalRestApiConfig.ts
@@ -12,9 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ContractAPI } from "@gala-chain/api";
-
-import { ContractConfig } from "../generic";
+import { ContractAPI, ContractConfig } from "@gala-chain/api";
 
 export interface RestApiAdminCredentials {
   adminKey: string;
